docsgen: move WriteCommand and WriteDefaultConfig to utils
This makes the functions importable by other projects.
